feat: 🎸 merge first_rows worker int datasets_based
