--- .github/workflows/w-first_rows-build-docker.yml
@@ -1,24 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 The HuggingFace Authors.
-
-name: workers/first_rows
-on:
-  workflow_dispatch:
-  push:
-    paths:
-      - 'workers/first_rows/Dockerfile'
-      - 'workers/first_rows/src/**'
-      - 'workers/first_rows/poetry.lock'
-      - 'workers/first_rows/pyproject.toml'
-      - '.github/workflows/w-first_rows-build-docker.yml'
-      - '.github/workflows/_build_push_docker_hub.yml'
-      - 'vendors/'
-jobs:
-  docker:
-    uses: ./.github/workflows/_build_push_docker_hub.yml
-    with:
-      directory: workers
-      project: first_rows
-    secrets:
-      dockerhub-username: ${{ secrets.DOCKERHUB_USERNAME }}
-      dockerhub-password: ${{ secrets.DOCKERHUB_PASSWORD }}
