--- .github/workflows/w-first_rows.yml
@@ -1,25 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 The HuggingFace Authors.
-
-name: workers/first_rows
-on:
-  workflow_dispatch:
-  push:
-    paths:
-      - 'workers/first_rows/**'
-      - '.github/workflows/w-first_rows.yml'
-      - '.github/workflows/_quality-python.yml'
-      - '.github/workflows/_unit-tests-python.yml'
-      - 'tools/docker-compose-mongo.yml'
-      - 'vendors/'
-jobs:
-  quality:
-    uses: ./.github/workflows/_quality-python.yml
-    with:
-      working-directory: workers/first_rows
-      is-datasets-worker: true
-  unit-tests:
-    uses: ./.github/workflows/_unit-tests-python.yml
-    with:
-      working-directory: workers/first_rows
-      is-datasets-worker: true
