+++ workers/datasets_based/README.md
@@ -12,6 +12,16 @@ The same worker is used for different endpoints to reuse shared code and depende
 
 - `DATASETS_BASED_ENDPOINT`: the endpoint on which the worker will work (pre-compute and cache the response). It can only be `/splits` at the moment.
 
+### First rows worker
+
+Only needed when the `DATASETS_BASED_ENDPOINT` is set to `/first-rows`: set environment variables to configure the first rows worker (`FIRST_ROWS_` prefix):
+
+- `FIRST_ROWS_FALLBACK_MAX_DATASET_SIZE`: the maximum size in bytes of the dataset to fallback in normal mode if streaming fails. Note that it requires to have the size in the info metadata. Set to `0` to disable the fallback. Defaults to `100_000_000`.
+- `FIRST_ROWS_MAX_BYTES`: the max size of the /first-rows endpoint response in bytes. Defaults to `1_000_000` (1 MB).
+- `FIRST_ROWS_MAX_NUMBER`: the max number of rows fetched by the worker for the split, and provided in the /first-rows endpoint response. Defaults to `100`.
+- `FIRST_ROWS_MIN_CELL_BYTES`: the minimum size in bytes of a cell when truncating the content of a row (see `FIRST_ROWS_ROWS_MAX_BYTES`). Below this limit, the cell content will not be truncated. Defaults to `100`.
+- `FIRST_ROWS_MIN_NUMBER`: the min number of rows fetched by the worker for the split, and provided in the /first-rows endpoint response. Defaults to `10`.
+
 ### Datasets library
 
 The following environment variables are used to configure two dependencies: the `datasets` and `numba` libraries:
