+++ workers/datasets_based/src/datasets_based/config.py
@@ -22,10 +22,28 @@ class DatasetsBasedConfig:
             self.processing_step = env.str(name="ENDPOINT", default="/splits")
 
 
+class FirstRowsConfig:
+    fallback_max_dataset_size: int
+    max_bytes: int
+    max_number: int
+    min_cell_bytes: int
+    min_number: int
+
+    def __init__(self):
+        env = Env(expand_vars=True)
+        with env.prefixed("FIRST_ROWS_"):
+            self.fallback_max_dataset_size = env.int(name="FALLBACK_MAX_DATASET_SIZE", default=100_000_000)
+            self.max_bytes = env.int(name="MAX_BYTES", default=1_000_000)
+            self.max_number = env.int(name="MAX_NUMBER", default=100)
+            self.min_cell_bytes = env.int(name="CELL_MIN_BYTES", default=100)
+            self.min_number = env.int(name="MIN_NUMBER", default=10)
+
+
 class AppConfig:
     cache: CacheConfig
     common: CommonConfig
     datasets_based: DatasetsBasedConfig
+    first_rows: FirstRowsConfig
     processing_graph: ProcessingGraphConfig
     queue: QueueConfig
     worker: WorkerConfig
@@ -35,6 +53,7 @@ class AppConfig:
         self.common = CommonConfig()
         self.cache = CacheConfig()
         self.datasets_based = DatasetsBasedConfig()
+        self.first_rows = FirstRowsConfig()
         self.processing_graph = ProcessingGraphConfig()
         self.queue = QueueConfig()
         self.worker = WorkerConfig()
