+++ workers/datasets_based/src/datasets_based/workers/first_rows.py
@@ -23,8 +23,8 @@ from libcommon.exceptions import CustomError
 from libcommon.utils import orjson_dumps
 from libcommon.worker import ConfigNotFoundError, SplitNotFoundError, Worker
 
-from first_rows.config import AppConfig, CacheConfig, FirstRowsConfig
-from first_rows.features import get_cell_value
+from datasets_based.config import AppConfig, CacheConfig, FirstRowsConfig
+from datasets_based.features import get_cell_value
 
 FirstRowsWorkerErrorCode = Literal[
     "SplitsNamesError",
@@ -558,7 +558,7 @@ class FirstRowsWorker(Worker):
             common_config=app_config.common,
             queue_config=app_config.queue,
             worker_config=app_config.worker,
-            version=importlib.metadata.version(__package__),
+            version=importlib.metadata.version(__package__.split(".")[0]),
         )
         self.cache_config = app_config.cache
         self.first_rows_config = app_config.first_rows
