--- workers/first_rows/.python-version
@@ -1,1 +1,0 @@
-3.9.6
