--- workers/first_rows/poetry.toml
@@ -1,3 +1,0 @@
-[virtualenvs]
-in-project = true
-prefer-active-python = true
