--- workers/first_rows/tests/conftest.py
@@ -1,32 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 The HuggingFace Authors.
-
-from pytest import MonkeyPatch, fixture
-
-from first_rows.config import AppConfig
-
-# Import fixture modules as plugins
-pytest_plugins = ["tests.fixtures.datasets", "tests.fixtures.files", "tests.fixtures.hub"]
-
-
-# see https://github.com/pytest-dev/pytest/issues/363#issuecomment-406536200
-@fixture(scope="session")
-def monkeypatch_session(hf_endpoint: str, hf_token: str):
-    monkeypatch_session = MonkeyPatch()
-    monkeypatch_session.setenv("CACHE_MONGO_DATABASE", "datasets_server_cache_test")
-    monkeypatch_session.setenv("QUEUE_MONGO_DATABASE", "datasets_server_queue_test")
-    monkeypatch_session.setenv("COMMON_HF_ENDPOINT", hf_endpoint)
-    monkeypatch_session.setenv("COMMON_HF_TOKEN", hf_token)
-    monkeypatch_session.setenv("COMMON_HF_TOKEN", hf_token)
-    monkeypatch_session.setenv("COMMON_ASSETS_BASE_URL", "http://localhost/assets")
-    monkeypatch_session.setenv("FIRST_ROWS_MAX_NUMBER", "7")
-    yield monkeypatch_session
-    monkeypatch_session.undo()
-
-
-@fixture(scope="session", autouse=True)
-def app_config(monkeypatch_session: MonkeyPatch) -> AppConfig:
-    app_config = AppConfig()
-    if "test" not in app_config.cache.mongo_database or "test" not in app_config.queue.mongo_database:
-        raise ValueError("Test must be launched on a test mongo database")
-    return app_config
