--- workers/first_rows/tests/fixtures/files.py
@@ -1,41 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright 2022 The HuggingFace Authors.
-
-import csv
-import json
-
-import pytest
-
-DATA = [
-    {"col_1": "0", "col_2": 0, "col_3": 0.0},
-    {"col_1": "1", "col_2": 1, "col_3": 1.0},
-    {"col_1": "2", "col_2": 2, "col_3": 2.0},
-    {"col_1": "3", "col_2": 3, "col_3": 3.0},
-]
-
-
-@pytest.fixture(scope="session")
-def csv_path(tmp_path_factory: pytest.TempPathFactory) -> str:
-    path = str(tmp_path_factory.mktemp("data") / "dataset.csv")
-    with open(path, "w", newline="") as f:
-        writer = csv.DictWriter(f, fieldnames=["col_1", "col_2", "col_3"])
-        writer.writeheader()
-        for item in DATA:
-            writer.writerow(item)
-    return path
-
-
-JSONL = [
-    {"col_1": "0", "col_2": 0, "col_3": 0.0},
-    {"col_1": None, "col_2": 1, "col_3": 1.0},
-    {"col_2": 2, "col_3": 2.0},
-    {"col_1": "3", "col_2": 3, "col_3": 3.0},
-]
-
-
-@pytest.fixture(scope="session")
-def jsonl_path(tmp_path_factory: pytest.TempPathFactory) -> str:
-    path = str(tmp_path_factory.mktemp("data") / "dataset.jsonl")
-    with open(path, "w", newline="") as f:
-        f.writelines(json.dumps(o) for o in JSONL)
-    return path
